* commit 'f87ce26e4961f336590a8af56ed00ffcc9d9b6fa':
  msm: camera: core: Reset last flush request id
  msm: camera: icp: Send cmd buffer memory region to FW
  msm: camera: smmu: Change shared mem pool granularity
  ARM: dts: msm: Add Camera sensor nodes to sdmmagpie
  ARM: dts: msm: Add CSI PHY and CCI nodes to sdmmagpie camera
  msm: camera: Changes to accommodate 32-bit Arch
  ARM: dts: msm: remove invalid clock rate in csid

+++ drivers/media/platform/msm/camera/cam_fd/fd_hw_mgr/fd_hw/cam_fd_hw_intf.h
@@ -111,7 +111,7 @@ struct cam_fd_hw_io_buffer {
 	struct cam_buf_io_cfg *io_cfg;
 	uint32_t               num_buf;
 	uint64_t               io_addr[CAM_PACKET_MAX_PLANES];
-	uint64_t               cpu_addr[CAM_PACKET_MAX_PLANES];
+	uintptr_t              cpu_addr[CAM_PACKET_MAX_PLANES];
 };
 
 /**
